fixed issue where tile can be multiple things
sometimes a space can be multiple
spawn, hero, or hero_adjacent. priorities have been set such that bugs
with tracking do not arrise

if a spawn is adjacent to a hero, it is treated as a spawn. if a hero
and spawn are coincident, it is treated as a hero. Eventually it would
be nice to consider every tile as just whatever we want to treat more
cautiously at that moment.